Fix a bug in the Ipc::Connection class
A defaulted move constructor/assignment operator would think that it
has unique ownership of the handle it's been given, leading to the
possibility of using a handle after it's been closed. This hasn't been
causing any issues due to NRVO kicking in in the only use case of the
class, but better to fix it now that it's been spotted.